refactor: lift shared infra utils to cube.infra_utils; move integration tests to harness
- Add cube/infra_utils.py: free_port, open_tunnel, open_tunnels, wait_for_ssh,
  BootstrapMonitor. Both _utils.py files become 4-line re-exports.
- Lift image_name_suffix, _image_name(), _resource_shim() onto InfraConfig base
  class in resource.py; update provision_status() to use _resource_shim() so
  the suffix is respected automatically without per-cloud overrides.
- Remove duplicate _image_name(), _resource_shim(), provision_status() from
  AWSInfraConfig and AzureInfraConfig (~60 lines removed per cloud).
- Remove benchmark cube dev deps (osworld-cube, webarena-verified-cube) from
  cube-infra-aws and cube-infra-azure pyproject.toml — infra packages must not
  depend on specific cubes.
- Delete test_run_debug_agent.py from both infra packages; integration tests
  now live in cube-harness/integration-tests/ with their own pyproject.toml.
