+++ cube-resources/cube-infra-aws/pyproject.toml
@@ -22,7 +22,6 @@ build-backend = "hatchling.build"
 [dependency-groups]
 dev = [
     "pytest>=9.0",
-    "osworld-cube",
 ]
 
 [tool.hatch.build.targets.wheel]
@@ -30,4 +29,3 @@ packages = ["src/cube_infra_aws"]
 
 [tool.uv.sources]
 cube-standard = { path = "../..", editable = true }
-osworld-cube = { path = "../../../cube-harness/cubes/osworld-cube", editable = true }
