+++ cube-resources/cube-infra-aws/src/cube_infra_aws/aws.py
@@ -325,11 +325,6 @@ class AWSInfraConfig(InfraConfig):
     bootstrap_role_name         str = "cube-bootstrap-role"
     bootstrap_profile_name      str = "cube-bootstrap"
 
-    ── Testing ───────────────────────────────────────────────────────────────────
-    image_name_suffix   str = ""
-        Suffix appended to AMI names and ProvisionStore keys.
-        Use "-test" to isolate test runs from production AMIs without touching
-        the production ProvisionStore entry.
     """
 
     # ── Auto-discovered ────────────────────────────────────────────────────────
@@ -355,9 +350,6 @@ class AWSInfraConfig(InfraConfig):
     bootstrap_root_volume_gb: int = 128
     bootstrap_role_name: str = "cube-bootstrap-role"
     bootstrap_profile_name: str = "cube-bootstrap"
-
-    # ── Testing ───────────────────────────────────────────────────────────────
-    image_name_suffix: str = ""
 
     # ── Auto-discovery ────────────────────────────────────────────────────────
 
@@ -463,16 +455,6 @@ class AWSInfraConfig(InfraConfig):
 
     # ── InfraConfig interface ─────────────────────────────────────────────────
 
-    def _image_name(self, resource: VMResourceConfig) -> str:
-        """AMI name (resource.name + image_name_suffix)."""
-        return resource.name + self.image_name_suffix
-
-    def _resource_shim(self, resource: VMResourceConfig) -> Any:
-        """Minimal object with .name = _image_name(resource) for ProvisionStore keys."""
-        import types
-
-        return types.SimpleNamespace(name=self._image_name(resource))
-
     def fingerprint(self) -> str:
         """Stable key: provider + region only (not instance type or S3 bucket).
 
@@ -483,16 +465,6 @@ class AWSInfraConfig(InfraConfig):
     def capabilities(self) -> set[str]:
         """EC2 HVM instances satisfy KVM requirements (same hypervisor technology)."""
         return {"kvm"}
-
-    def provision_status(self, resource: ResourceConfig) -> Literal["ready", "needs_provisioning"]:
-        """Query ProvisionStore using the effective image name (respects image_name_suffix)."""
-        from cube.provision_store import ProvisionStore
-
-        if not isinstance(resource, VMResourceConfig):
-            return "needs_provisioning"
-        shim = self._resource_shim(resource)
-        store = ProvisionStore()
-        return "ready" if store.get(shim, self) is not None else "needs_provisioning"
 
     def provision(self, resource: ResourceConfig) -> None:
         """Bootstrap a VM image from source_url into an EC2 AMI.
